✨(scripts) add build & publish scripts
Add bash scripts to automate service image builds and publication.

* `build`: usage: `bin/build SERVICE`

Check that a Dockerfile exists for this service (it is expected to be in
`docker/images/<SERVICE>/Dockerfile`), get the service tag and build an
image tagged with our dockerhub namespace (fundocker), an image prefix
(openshift), the service name and the original image tag, _e.g._

fundocker/openshift-nginx:1.13

* `publish`: usage: `bin/publish SERVICE`

If the current release of the service image has been build, push it to
DockerHub.