👷(circle) add build and publication jobs
We always build all services to ensure nothing is broken in a PR. A
service is published when the git repository has been tagged with the
following pattern: <service>-<version> (e.g. nginx-1.13).

To publish all services, we need to tag the git repository with the
following pattern: all-<date> (e.g. all-20180420).